[Resolver] Refactored Resolvers and wrote tutorial.
