+++ doc/source/examples.rst
@@ -27,7 +27,7 @@ Misc
     notebooks/bullet_world
     notebooks/minimal_task_tree
     notebooks/pose
-    notebooks/custom_resolver
+    notebooks/improving_actions
     notebooks/language
     notebooks/local_transformer
